requirements.txt
attr==0.3.2
ConfigParser==6.0.0
cryptography==42.0.2
Cython==3.0.7
Django==4.2.7
dl==0.1.0
docutils==0.20.1
HTMLParser==0.0.2
ipython==8.15.0
ipywidgets==8.1.2
Jinja2==3.1.2
keyring==24.3.0
lockfile==0.12.2
numpy==1.26.4
Pillow==10.2.0
protobuf==4.25.2
pyOpenSSL==24.0.0
railroad==0.5.0
Sphinx==7.2.6
thread==0.1.3
tornado==6.4
trove_classifiers==2024.1.31
truststore==0.8.0
urllib3_secure_extra==0.1.0
djangorestframework==3.14.0
